NSA-386 Implemented models for get and post functionality

+++ lib/models/size/request/size_tracking_request_model.dart
@@ -15,14 +15,14 @@ class SizeTrackingRequestModel {
         this.comment,
     });
 
-    int chest;
-    int leftArm;
-    int rightArm;
-    int abdomen;
-    int waist;
-    int hips;
-    int leftThigh;
-    int rightThigh;
+    double chest;
+    double leftArm;
+    double rightArm;
+    double abdomen;
+    double waist;
+    double hips;
+    double leftThigh;
+    double rightThigh;
     String? comment;
     int measureUnitId;
     DateTime registered;
